run bt

+++ src/main/java/com/yoson/task/BackTestTask.java
@@ -163,20 +163,22 @@ public class BackTestTask implements Runnable {
 		}
 		BackTestCSVWriter.writeText(mainUIParam.getParamPath(), new Gson().toJson(mainUIParam), false);
 		int startStep = 0;
-		try {
-			File stepFile = new File(mainUIParam.getStepPath());
-			if(stepFile.exists()) {
-				FileInputStream input = new FileInputStream(stepFile);
-				String step = IOUtils.toString(input);
-				input.close();
-				if(StringUtils.isNotEmpty(step)) {
-					startStep = Integer.parseInt(step.split(",")[0]);
-				}
-			}
-		} catch (Exception e) {
+		if (!isLiveData) {
 			try {
-				FileUtils.cleanDirectory(new File(mainUIParam.getSourcePath()));			
-			} catch (Exception ex) {
+				File stepFile = new File(mainUIParam.getStepPath());
+				if(stepFile.exists()) {
+					FileInputStream input = new FileInputStream(stepFile);
+					String step = IOUtils.toString(input);
+					input.close();
+					if(StringUtils.isNotEmpty(step)) {
+						startStep = Integer.parseInt(step.split(",")[0]);
+					}
+				}
+			} catch (Exception e) {
+				try {
+					FileUtils.cleanDirectory(new File(mainUIParam.getSourcePath()));			
+				} catch (Exception ex) {
+				}
 			}
 		}
 		boolean first = (startStep == 0);
@@ -228,22 +230,22 @@ public class BackTestTask implements Runnable {
 			if (!BackTestTask.running)
 				break;
 		}
+		Set<Integer> specifyDateRanges = new TreeSet<Integer>();
+		if (!StringUtils.isBlank(mainUIParam.getnForPnl())) {
+			for(String nForPnl : mainUIParam.getnForPnl().split(",")) {
+				try {
+					specifyDateRanges.add(Integer.parseInt(nForPnl));				
+				} catch (Exception e) {
+				}
+			}			
+		}
+		BackTestCSVWriter.writeText(FilenameUtils.concat(mainUIParam.getSourcePath(), BackTestCSVWriter.profitAndLossByDateFileName), BackTestCSVWriter.getBestPnlByDate(), true);
+		BackTestCSVWriter.writeText(FilenameUtils.concat(mainUIParam.getSourcePath(), BackTestCSVWriter.profitAndLossByDateRangeFileName), BackTestCSVWriter.getBestPnlBySpecifyDates(specifyDateRanges), true);
 		milliseconds = System.currentTimeMillis() - start;
 		callBack.updateStatus(getStatus("All task done, total time cost: " + DateUtils.dateDiff(milliseconds)));
 		
 //		BackTestCSVWriter.writeCSVResult(mainUIParam);
 		if (!isLiveData) {
-			Set<Integer> specifyDateRanges = new TreeSet<Integer>();
-			if (!StringUtils.isBlank(mainUIParam.getnForPnl())) {
-				for(String nForPnl : mainUIParam.getnForPnl().split(",")) {
-					try {
-						specifyDateRanges.add(Integer.parseInt(nForPnl));				
-					} catch (Exception e) {
-					}
-				}			
-			}
-			BackTestCSVWriter.writeText(FilenameUtils.concat(mainUIParam.getSourcePath(), BackTestCSVWriter.profitAndLossByDateFileName), BackTestCSVWriter.getBestPnlByDate(), true);
-			BackTestCSVWriter.writeText(FilenameUtils.concat(mainUIParam.getSourcePath(), BackTestCSVWriter.profitAndLossByDateRangeFileName), BackTestCSVWriter.getBestPnlBySpecifyDates(specifyDateRanges), true);
 			SQLUtils.saveTestSetResult(FilenameUtils.concat(mainUIParam.getSourcePath(), BackTestCSVWriter.btSummaryFileName), mainUIParam.getVersion().replaceAll(" ", ""));
 		}
 		gc();
